Remove import results page
