--- app/views/admin/_import_results_steps.html.erb
@@ -1,89 +1,0 @@
-<div id="result-inbox-steps">
-  <% if @existing_data.values.all?(&:zero?) %>
-    <%= alert :info do %>
-      This competition has no data yet. <%= link_to "Upload something", competition_admin_upload_results_edit_path(@competition.id) %> to get started.
-    <% end %>
-  <% else %>
-    <% if @inbox_step.present? %>
-      <%= alert :info do %>
-        <p>This competition is in the process of having data uploaded. You
-          can <%= link_to "run validators", competition_admin_check_existing_results_path(@competition.id), target: '_blank' %>
-          at any time, but you will need to refresh this page manually afterwards.</p>
-      <% end %>
-    <% end %>
-
-    <div class="progress">
-      <% inbox_step_num = (AdminController::RESULTS_POSTING_STEPS.index(@inbox_step) || AdminController::RESULTS_POSTING_STEPS.length) + 1 %>
-      <% number_of_steps = AdminController::RESULTS_POSTING_STEPS.length + 1 # count Sanity Checks as additional step %>
-
-      <% progress_percent = inbox_step_num.to_f / number_of_steps * 100 %>
-
-      <% results_public = @competition.results_posted? %>
-
-      <div class="progress-bar <%= "progress-bar-success" if results_public %>" role="progressbar" style="width: <%= progress_percent %>%">
-        <% if @inbox_step.present? %>
-          Step <%= inbox_step_num %>/<%= number_of_steps %>
-        <% elsif !results_public %>
-          Import complete, but not published yet…
-        <% else %>
-          All steps completed and published!
-      <% end %>
-      </div>
-    </div>
-
-    <% if @inbox_step.present? %>
-      <% AdminController::RESULTS_POSTING_STEPS.each do |step| %>
-        <%= render 'import_inbox_step', render_step: step %>
-      <% end %>
-    <% else %>
-      <p>All inbox data has been imported. You should consider the following steps next:</p>
-      <ul>
-        <li><%= link_to "Check record markers", admin_override_regional_records_path(check_regional_records_form: { competition_id: @competition.id, event_id: 'all', refresh_index: true }), target: '_blank' %></li>
-        <li><%= link_to "Compute auxiliary data", admin_do_compute_auxiliary_data_path, target: '_blank' %></li>
-      </ul>
-      <p>
-        You can also visit the <%= link_to "public results page", competition_results_all_path(@competition, event: 'all'), target: '_blank' %> as an additional sanity check.
-        Once you are sure, you can post the results using the button below:
-      </p>
-      <%= button_to competition_post_results_path(@competition), class: 'btn btn-primary', disabled: @competition.results_posted?, method: :post, data: { confirm: 'You are about to publish the results, including sending out emails to the competitors. Are you sure?'} do %>
-        <%= t 'competitions.post_results' %>
-      <% end %>
-      <% if @competition.results_posted? %>
-          <p class="help-block">
-            <%= t 'competitions.results_posted_by_html', poster_name: User.find(@competition.results_posted_by).name, date_time: wca_local_time(@competition.results_posted_at) %>
-          </p>
-      <% end %>
-    <% end %>
-  <% end %>
-
-  <script>
-    $("a.inbox-trigger").on('click', function(e) {
-      e.preventDefault();
-      var $el = $(this);
-
-      var httpMethod = $el.data('jquery-method') || 'GET';
-
-      var confirmationMessage = $el.data('jquery-confirm');
-      var userConfirms = !confirmationMessage || confirm(confirmationMessage);
-
-      if (userConfirms) {
-        var progressBar = $('div.progress-bar')
-        progressBar.removeClass('progress-bar-danger').addClass('progress-bar-striped active');
-
-        window.wca.cancelPendingAjaxAndAjax('inbox-trigger', {
-          url: $el.attr('href'),
-          method: httpMethod.toUpperCase(),
-          success: function(html) {
-            $('#result-inbox-steps').replaceWith(html);
-          },
-          error: function(response) {
-            progressBar.removeClass('progress-bar-success active').addClass('progress-bar-danger');
-
-            var requestId = response.getResponseHeader('X-Request-Id');
-            progressBar.text(`An error occured! Request ID ${requestId}`)
-          }
-        });
-      }
-    });
-  </script>
-</div>
