+++ app/views/competitions/_nav.html.erb
@@ -104,7 +104,6 @@
             icon: "cogs",
             children: [
               { text: "Upload and check results", path: competition_admin_upload_results_edit_path(@competition), icon: "check" },
-              { text: "Import uploaded results", path: competition_admin_import_results_path(@competition), icon: "check" },
               {
                 text: "Check existing results",
                 path: competition_admin_check_existing_results_path(@competition),
