+++ app/webpacker/components/Tickets/TicketWorkbenches/CompetitionResultActionerView/index.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { DateTime } from 'luxon';
 import { ticketsCompetitionResultStatuses } from '../../../../lib/wca-data.js.erb';
 import WarningsVerification from './WarningsVerification';
 import TimelineView from './TimelineView';
@@ -6,6 +7,7 @@ import MergeInboxResults from './MergeInboxResults';
 import CreateWcaIds from './CreateWcaIds';
 import FinalSteps from './FinalSteps';
 import MiscActions from './MiscActions';
+import I18n from '../../../../lib/i18n';
 
 export default function CompetitionResultActionerView({ ticketDetails, currentStakeholder }) {
   const { ticket: { metadata: { status } } } = ticketDetails;
@@ -28,6 +30,14 @@ export default function CompetitionResultActionerView({ ticketDetails, currentSt
 function ViewForStatus({
   status, ticketDetails, currentStakeholder,
 }) {
+  const {
+    ticket: {
+      metadata: {
+        competition: { results_posted_at: resultsPostedAt, posted_user: postedUser },
+      },
+    },
+  } = ticketDetails;
+
   switch (status) {
     case ticketsCompetitionResultStatuses.submitted:
       return <p>Please lock the competition results from the Posting dashboard.</p>;
@@ -59,6 +69,15 @@ function ViewForStatus({
         <FinalSteps
           ticketDetails={ticketDetails}
         />
+      );
+    case ticketsCompetitionResultStatuses.posted:
+      return (
+        <>
+          {I18n.t('competitions.results_posted_by_html', {
+            poster_name: postedUser.name,
+            date_time: DateTime.fromISO(resultsPostedAt).toLocaleString(DateTime.DATETIME_FULL),
+          })}
+        </>
       );
 
     default:
