+++ config/locales/en.yml
@@ -1363,7 +1363,6 @@ en:
       results_not_posted: "This competition's results are visible but haven't been posted yet."
       results_still_processing: "We are busy processing this competition's results - they should be available shortly."
       results_already_posted: "The results have already been posted."
-      results_posted: "Successfully posted results!"
       results_preview_alert: "You are viewing results that have not been posted yet."
       upload_results: "This competition is over. We are working to upload the results as soon as possible!"
       computing_auxiliary_data: "Please wait until auxiliary data is computed."
@@ -2074,7 +2073,6 @@ en:
     uncancel: "Uncancel the competition"
     cancelled_by_html: "Cancelled by %{name} on %{date_time}"
     cancel_confirm: "Are you sure you want to cancel that competition?"
-    post_results: "Post the results"
     results_posted_by_html: "Posted by %{poster_name} on %{date_time}"
     cancel_mistake: "Made a mistake?"
     note_before_cancel: "Note: only confirmed competitions can be cancelled."
